% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{data}
\name{docs}
\alias{docs}
\title{docs}
\format{
A data frame with 3 variables: \code{key}, \code{value_type} and
\code{key_description}
}
\usage{
docs
}
\description{
Source: \href{https://eerscmap.usgs.gov/uspvdb/api-doc/}{USPVDB Website}
}
\details{
A dataset containing the decription of the api response from the uspvdb api.

Values in the \code{value_type} column have been converted to valid R types.
}
\keyword{datasets}
