% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\title{Parse API response}
\usage{
parse(x, data_type = "json")
}
\arguments{
\item{x}{A response object as created by calling 'get_*' function}

\item{data_type}{The type of data to parse. Default is 'json'. All documented
responses are JSON.}
}
\value{
tibble when data_type is json
}
\description{
Parse API response
}
\examples{
\dontrun{
x <- get_("projects")
parse(x)
}
}
