% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{perform}
\alias{perform}
\title{Perform a request}
\usage{
perform(call, debug = FALSE, verbosity = 3)
}
\arguments{
\item{call}{A request object as created by 'httr2::request()'}

\item{debug}{If TRUE, the request is not actually performed, but a dry run is
shown instead. This is useful in debugging. Default is FALSE.}

\item{verbosity}{The level of information to print to the users console. See
?httr2::req_perform for details. Default is 3.}
}
\value{
A response object
}
\description{
Intended to be used internal to the package.
}
\examples{
\dontrun{
req <- httr2::request("https://httpbin.org/get")
perform(req, TRUE)
perform(req)
}
}
