% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uspvdb.R
\docType{package}
\name{uspvdb-package}
\alias{uspvdb}
\alias{uspvdb-package}
\title{uspvdb: Helper functions to interact with the USPVDB API}
\description{
TODO - To learn more about uspvdb, start with the vignette:
\code{browseVignettes(package = "uspvdb")}
}
\author{
\strong{Maintainer}: Josh Nowak \email{josh.nowak@speedgoat.io} (\href{https://orcid.org/0000-0002-8553-7450}{ORCID})

Other contributors:
\itemize{
  \item SpeedGoat [copyright holder]
}

}
